requirements.txt
Django==1.8.3
djangular
django-angular
djangorestframework
markdown
django-filter
mysqlclient